Deploy with curl instead
I initially switched back to netlify-cli because I thought we're
deploying a ~200MB build and could benefit from partial uploads. No,
the 200MB is for the full build, the one with 170000+ files that
Netlify (reasonably) rejects, and since we build on first request the
build time deploy should be quite small.

This allows us to not have to install netlify-cli.