Use updated dict_revised data
Instead of using moedict-data, this is now using the lightly
processed and updated dict_revised data that I'm mirroring at
https://github.com/kisaragi-hiu/kemdict-data-ministry-of-education.

The processing involves converting XLSX to CSV then to JSON, along
with a header rename. It's much lighter than what moedict-data does;
this also means it needs to be processed further before it can be used
in a template.

This is partly because I can't / don't want to figure out how to
update moedict-data to the latest version (20220922). It is still
stuck on version 20160523.