Stop building for Netlify; don't build data on exact cache hit
I thought Make would handle determining whether the data should be
built or not for us, but it can't because Git doesn't restore
metadata. So we have to skip it ourselves.

The cache hit is only exact when nothing in dicts/ has
changed (because of the hashFiles() call), so this will work fine.