put kisaragi-dict up top
If there's a word that has both definitions from an MOE dictionary and
kisaragi-dict, that means I'm trying to add something to the
definition because there is something missing. Given that's already
what I'm trying to do, it's worth it to just put it on top so it's
actually seen.