fix empty strings being in the pronunciations list
The omit-nulls argument doesn't omit strings that would become empty
after trimming. You can get it to do that by using split-string
directly and passing in the TRIM argument, but the TRIM argument has
to be given a regexp and there's no default that is readily available,
unlike string-trim.

Doing it this way gives us cleaner code IMO.