Avoid 8th tone i in POJ displaying weirdly
The dot and the 8th tone COMBINING VERTICAL LINE ABOVE are supposed to
fuse, but for some reason when loading from a webfont (woff2, tried
Google Fonts and @fontsource) it just doesn't. Loading the TTF is an
option, but that causes seconds of blank text.

Try to use system fonts instead.