Update dicts
Fix pronunciations being normalized to NFC, and not NFD like the rest
of the database. The result was that rendering multiple words with the
same pronunciation in the same word page no longer worked if the
pronunciation has an accent, or more accurately if its NFC and NFD
forms differed.