Remove locally installed eslint
This is -72 packages.

The JS world apparently wants every single project to pin every
development tool they use and install them locally. If pnpm isn't
used, and every project follows this "best practice", every project
will have its own instance of every editing tool that it needs.

This is ridiculous.

An editing tool doesn't need to be versioned as precisely as a library
dependency. Pinning them makes it more convenient for a new
developer to the repository to set up their development environment,
but getting every project to install the same tools over and over
again isn't the best way, or even a good way. It's standardized and
arguably works, but it is still ridiculous.

Install scripts and helper scripts might not work as well, but we've
settled on a dirty workaround without even remembering how ridiculous
it is.

If using a linter is depending on a linter, literally every project
that follows the best practice will have to depend on a linter, and
this is what JS best practice demands for every project right now. It
works fine, across platforms and across familiarities with the
ecosystem, but it's still a dirty mess.

An ecosystem-wide linter should hopefully eventually settle down such
that it has a core that is safe to install globally and assume to be
unchanging.

I would also like to remove locally installed prettier, but

- The current version of astro (2.10.3) depends on
  @astrojs/language-server ^1.0.0 (???)
- @astrojs/language-server 1.0.8 depends on prettier 2.8.8
- so `npx prettier` is always 2.8.8

It's a mess, and I honestly am not sure what it should've been like.