fix build error when DB isn't present
Previously, importing db.ts would immediately calculate `chars`, which
depends on the DB being present. This wasn't a problem because
presumably the file wasn't being imported during build.

This has changed, so we explicitly make it initialize the first time
the new function getGroupedChars is called instead.

We also make db.ts look less hard for the DB, so that it's easier to
change the development environment to one where the DB isn't present,
while still keeping the DB in the original place where it was built.