workaround entries with an empty title
dict_revised appears to have multiple entries with an empty title. You
can check them out by searching for "title":"" (including the quotes) in
dict_revised.json.

I'm in the process of figuring out if they are an artifact of xlsx->json
conversion or if they are indeed empty in the original spreadsheet. For
now, we should just not add those entries.